Updated with new UUIDs
I'm going to have to keep doing this. I'm not sure if there's a better way to do this, but I need to figure it out. I may just do a get request and then download the json, then parse it into the correct format for swift. seems stupid but I'm not sure if there's a better way. If I can write a script then this could be a pretty automated process despite whatever DB changes I make, I could just rerun the script.