Added HomeCardView
I added HomeCardView and made it somewhat customizable. I have some commented out code if i want to abstract less to the card. I'm not sure but im gonna push this commit and then make another commit removing that code (just for clean up purposes). Refer to this commit for code that achieves this (or for the code without the card)